// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface ServiceDocumentDataPricingPlansItem {
  /**
   * plan_description field in *service → Pricing plans*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: service.pricing_plans[].plan_description
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  plan_description: prismic.RichTextField;

  /**
   * plan_name field in *service → Pricing plans*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: service.pricing_plans[].plan_name
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  plan_name: prismic.KeyTextField;

  /**
   * timeline field in *service → Pricing plans*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: service.pricing_plans[].timeline
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  timeline: prismic.KeyTextField;

  /**
   * features field in *service → Pricing plans*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: service.pricing_plans[].features
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  features: prismic.RichTextField;

  /**
   * button field in *service → Pricing plans*
   *
   * - **Field Type**: Link
   * - **Placeholder**: *None*
   * - **API ID Path**: service.pricing_plans[].button
   * - **Documentation**: https://prismic.io/docs/fields/link
   */
  button: prismic.LinkField<string, string, unknown, prismic.FieldState, never>;

  /**
   * featured field in *service → Pricing plans*
   *
   * - **Field Type**: Boolean
   * - **Placeholder**: *None*
   * - **Default Value**: false
   * - **API ID Path**: service.pricing_plans[].featured
   * - **Documentation**: https://prismic.io/docs/fields/boolean
   */
  featured: prismic.BooleanField;

  /**
   * price field in *service → Pricing plans*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: service.pricing_plans[].price
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  price: prismic.KeyTextField;
}
// ...
interface ServiceDocumentData {
// ...
  pricing_plans: prismic.GroupField<
    Simplify<ServiceDocumentDataPricingPlansItem>
  >;
// ...
}

/**
 * service document from Prismic
 *
 * - **API ID**: `service`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type ServiceDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<ServiceDocumentData>,
    "service",
    Lang
  >;